refactor: SearchSection 구조 리팩토링 #21
