+++ src/components/sections/SearchSection/SearchSection.css
