+++ src/pages/SearchResult.jsx
@@ -1,0 +1,11 @@
+import SearchSection from "../components/sections/SearchSection/SearchSection";
+
+const SearchResult = () => {
+  return (
+    <div>
+      <SearchSection />
+    </div>
+  );
+};
+
+export default SearchResult;
